+++ app/src/main/java/com/example/soundsensev1/DrawerItem.java
@@ -15,7 +15,7 @@ public abstract class DrawerItem<T extends DrawerAdapter.ViewHolder> {
     }
 
     public boolean isChecked(){
-        return isChecked();
+        return isChecked;
     }
 
     public boolean isSelectable(){
