+++ app/src/main/java/com/example/soundsensev1/SimpleItem.java
@@ -37,8 +37,8 @@ public class SimpleItem extends DrawerItem<SimpleItem.ViewHolder>{
         holder.title.setText(title);
         holder.icon.setImageDrawable(icon);
 
-        holder.title.setText(isChecked ? selectedItemTextTint : normalItemTextTint);
-        holder.icon.setColorFilter(isChecked ? selectedItemIconTint : normalItemIconTint);
+        //holder.title.setText(isChecked() ? selectedItemTextTint : normalItemTextTint);
+        //holder.icon.setColorFilter(isChecked() ? selectedItemIconTint : normalItemIconTint);
     }
 
     public SimpleItem withSelectedIconTint(int selectedItemIconTint){
