--- app/src/main/res/layout/activity_data.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@mipmap/backgound1"
-    tools:context=".DataActivity">
-
-    <include
-        android:id="@+id/mainToolbar"
-        layout="@layout/toolbar"
-        android:layout_width="411dp"
-        android:layout_height="50dp" />
-
-    <ListView
-        android:id="@+id/sensorListView"
-        android:layout_width="339dp"
-        android:layout_height="505dp"
-        android:layout_marginStart="38dp"
-        android:layout_marginTop="40dp"
-        android:layout_marginEnd="38dp"
-        android:layout_marginBottom="24dp"
-        app:layout_constraintBottom_toTopOf="@+id/deleteListButton"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/mainToolbar" />
-
-    <Button
-        android:id="@+id/deleteListButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="143dp"
-        android:layout_marginEnd="143dp"
-        android:layout_marginBottom="47dp"
-        android:text="Delete List"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
