+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:alpha="1"
-    android:background="@mipmap/backgound1"
+    android:background="@color/white"
     tools:context=".MainActivity">
 
 
@@ -15,44 +15,11 @@
         tools:layout_editor_absoluteX="0dp"
         tools:layout_editor_absoluteY="1dp" />
 
+
     <FrameLayout
         android:id="@+id/container"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_weight="1"/>
-
-    <TextView
-        android:id="@+id/welcomeTextView"
-        android:layout_width="0dp"
-        android:layout_height="53dp"
-        android:layout_marginTop="128dp"
-        android:layout_marginBottom="20dp"
-        android:textAlignment="center"
-        android:textColor="@color/white"
-        android:textSize="24sp"
-        app:layout_constraintBottom_toTopOf="@+id/mainButton"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/mainToolbar" />
-
-    <Button
-        android:id="@+id/mainButton"
-        android:layout_width="195dp"
-        android:layout_height="197dp"
-        android:layout_marginStart="108dp"
-        android:layout_marginTop="20dp"
-        android:layout_marginEnd="108dp"
-        android:layout_marginBottom="267dp"
-        android:background="@drawable/circular_button"
-        android:textAlignment="center"
-        android:textAllCaps="true"
-        android:textColor="@color/white"
-        android:textSize="20sp"
-        app:backgroundTint="@null"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/welcomeTextView" />
-
 
 </androidx.constraintlayout.widget.ConstraintLayout>
