--- app/src/main/res/layout/activity_settings.xml
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@mipmap/backgound1"
-    tools:context=".SettingsActivity">
-
-    <include
-        android:id="@+id/mainToolbar"
-        layout="@layout/toolbar"
-        android:layout_width="392dp"
-        android:layout_height="42dp"
-        tools:layout_editor_absoluteX="10dp" />
-
-    <Switch
-        android:id="@+id/controlSwitch"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="60dp"
-        android:layout_marginEnd="61dp"
-        android:layout_marginBottom="118dp"
-        app:layout_constraintBottom_toTopOf="@+id/controlSwitch2"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/mainToolbar" />
-
-    <Switch
-        android:id="@+id/controlSwitch2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="61dp"
-        android:layout_marginBottom="457dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="225dp"
-        android:layout_height="26dp"
-        android:layout_marginStart="30dp"
-        android:layout_marginTop="60dp"
-        android:layout_marginEnd="49dp"
-        android:layout_marginBottom="119dp"
-        android:text="Device Control ON/OFF"
-        android:textColor="@color/white"
-        android:textSize="20sp"
-        app:layout_constraintBottom_toTopOf="@+id/textView3"
-        app:layout_constraintEnd_toStartOf="@+id/controlSwitch"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/mainToolbar" />
-
-    <TextView
-        android:id="@+id/textView3"
-        android:layout_width="225dp"
-        android:layout_height="26dp"
-        android:layout_marginStart="30dp"
-        android:layout_marginEnd="49dp"
-        android:layout_marginBottom="458dp"
-        android:text="Notifications ON/OFF"
-        android:textColor="#FAF9F9"
-        android:textSize="20sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/controlSwitch2"
-        app:layout_constraintStart_toStartOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
